+++ NEWS.md
@@ -15,6 +15,7 @@ New features(CLI, Configs):
 + Properly render the progress bar when Phan runs with multiple processes (#2928)
 + Add an HTML output mode to generate an unstyled HTML fragment.
   Example CSS styles can be generated with `internal/dump_html_styles.php`
++ Add a `light` color scheme for white backgrounds.
 
 New features(Analysis):
 + Fix failure to infer real types when an invoked function or method had a phpdoc `@return` in addition to the real type.